Remove tabulate dependency, change of format
Removed tabulate dependecy from `run` script. Now all external
dependencies are removed from the script, it is now possible to run
the script from DiffKemp's `nix` development environment.

The main change is that the output file of `run` script was
renamed from `result.md` to `result.json`. `tabulate` function is not
anymore used for creating the file. The file is now in JSON format.
The format with the description of the fields is documented in
`README.md`.

Changes:
- The results are not calculated from the `eqbench-results.csv` file,
  but are aggregated inside the program.
- Instead of using naming like `Eq correct` is used naming `TN` meaning
  true negative.
- The `result.json` file also contains total results.
- The stdout output is different from the `result.json` file and
  contains only total results.
- `analysis` script was edited to be able to parse results
  from `result.json` file.