Remove yaml dependency
Removed yaml dependecy from `run` script.

To remove the dependency it was necessary to change the way
how aggregated functions are evaluated. Previously all functions
were compared and then was checked if all functions (eliminating
the ones which should not be compared) are equal/non-equal. For
the elimination was used the `yaml` dependency so this had to be
changed. Now we specify list of programs, for which should be compared
only ceratin functions, when building snapshots of the programs
on the list we build the snapshots with only these functions so we
can later compare all functions.