Add option for adding options to compare command
It is now possible to pass additional options to compare
command by using `--add-cmp-opt`. This can be useful because
DiffKemp can contain new options (e.g. new extensions which are
not enabled by default) and this allows to use them without
need to edit the script.

To make sure that the option is valid, check, if the options
is present in help of the compare command, was added.
The check is run before building/comparing the programs.
The check is only simple, so it will not handle everything.